template sudy v7

+++ R/createStudy.R
@@ -10,6 +10,7 @@
 #' @export
 #' 
 #' @importFrom whisker whisker.render
+#' @importFrom utils unzip
 #'
 #' @examples
 #' \dontrun{
@@ -17,9 +18,7 @@
 #' createStudy("path/to/simulation")
 #' 
 #' }
-createStudy <- function(path, study_name = "my_study", antares_version = "6.0.0") {
-  if (antares_version >= "6.5.0") 
-    warning("You may need to open the study in Antares before modifying it!", call. = FALSE)
+createStudy <- function(path, study_name = "my_study", antares_version = "7.0.0") {
   if (!dir.exists(path)) {
     dir.create(path = path, recursive = TRUE)
   } else {
@@ -27,12 +26,16 @@ createStudy <- function(path, study_name = "my_study", antares_version = "6.0.0"
       path <- file.path(path, study_name)
     }
   }
-  statut <- file.copy(
-    from = list.files(path = system.file("newStudy", package = "antaresEditObject"), full.names = TRUE),
-    to = path, recursive = TRUE
-  )
-  to_delete <- list.files(path = path, pattern = "ANTARESEDITOBJECT_TODELETE", full.names = TRUE, recursive = TRUE)
-  unlink(to_delete)
+  if (antares_version < "6.5.0") {
+    statut <- file.copy(
+      from = list.files(path = system.file("newStudy", package = "antaresEditObject"), full.names = TRUE),
+      to = path, recursive = TRUE
+    )
+    to_delete <- list.files(path = path, pattern = "ANTARESEDITOBJECT_TODELETE", full.names = TRUE, recursive = TRUE)
+    unlink(to_delete)
+  } else {
+    statut <- unzip(zipfile = system.file("template-antares/antares-study-v7.zip", package = "antaresEditObject"), exdir = path)
+  }
   antares <- paste(readLines(con = file.path(path, "study.antares")), collapse = "\n")
   antares <- whisker::whisker.render(
     template = antares,
